added redux to city instead filters
