+++ project/src/components/main/city.tsx
@@ -1,18 +1,35 @@
+import { connect, ConnectedProps } from 'react-redux';
 import { Link } from 'react-router-dom';
+import { Dispatch } from 'redux';
+import { changeCity } from '../../store/action';
+import { Actions, State } from '../../types/reducer';
 
 type CityProps = {
   town: string,
-  onClick: (value: string) => void,
-  currentValue: string,
 }
 
-function City({ town, onClick, currentValue = 'Paris' }: CityProps): JSX.Element {
+const mapStateToProps = ({ city }: State) => ({
+  city,
+});
+
+const mapDispatchToProps = (dispatch: Dispatch<Actions>) => ({
+  onChangeCity(city: string) {
+    dispatch(changeCity(city));
+  },
+});
+
+const connector = connect(mapStateToProps, mapDispatchToProps);
+
+type PropsFromRedux = ConnectedProps<typeof connector>;
+type ConnectedComponentProps = PropsFromRedux & CityProps;
+
+function City({ town, onChangeCity, city = 'Paris' }: ConnectedComponentProps): JSX.Element {
   return (
     <li className="locations__item">
       <Link
-        className={ `locations__item-link tabs__item ${ currentValue === town && 'tabs__item--active'}` }
+        className={ `locations__item-link tabs__item ${ city === town && 'tabs__item--active'}` }
         to="/"
-        onClick={ () => onClick(town) }
+        onClick={ () => onChangeCity(town) }
       >
         <span>{ town }</span>
       </Link>
@@ -20,4 +37,5 @@ function City({ town, onClick, currentValue = 'Paris' }: CityProps): JSX.Element
   );
 }
 
-export default City;
+export { City };
+export default connector(City);
